Move SceneryGroupEntry
