+++ src/openrct2/libopenrct2.vcxproj
@@ -319,6 +319,7 @@
     <ClInclude Include="object\ObjectType.h" />
     <ClInclude Include="object\ResourceTable.h" />
     <ClInclude Include="object\RideObject.h" />
+    <ClInclude Include="object\SceneryGroupEntry.h" />
     <ClInclude Include="object\SceneryGroupObject.h" />
     <ClInclude Include="object\SceneryObject.h" />
     <ClInclude Include="object\SmallSceneryEntry.h" />
